chore: Separating files and adding second constructor for ElementMetadataUpdateHandler
